Create a new bun reimplementation
Reimplement bun to take care of NxN mapping.
